refactor: use relative standalone compat consumers
